Relocate IPFS client and peering helpers
refactor(network): move Kubo client to clients/ipfs_client.py;
 rename ipfs_docker to docker_ipfs_transport_peering;
 update imports and docs paths
